Web Crawler
Python based project that will input any URL of the website and start
crawling over it getting back with uselfull information.
The information can be the list of all the hyperlinks used on a web
page, then it will crawl to next hyperlink for the same process. The
next version of the project will download images and videos from the web
page storing it into the same folder where your project is.
Apart from this a concept of security breaker is also used. Which by
pass some small level security that dosent allow a programe to get
source code.